frontend

+++ backend/controllers/product.controller.js
@@ -42,6 +42,9 @@ export const updateProduct = async (req,res) => {
 }
 export const deleteProduct = async (req,res) => {
     const {id} = req.params;
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(404).json({success:false, message: "Id not found"});
+    }
     try{
         await Product.findByIdAndDelete(id);
         res.status(200).json({success:true, message:"Product deleted"});
